+++ include/dsn/tool-api/global_config.h
@@ -33,6 +33,10 @@
  *     xxxx-xx-xx, author, fix bug about xxx
  */
 
+/// Attention: There are some types which are defined in dsn_runtime being used in this file, so
+/// this file is coupled with dsn_runtime. So if you want to add some variables/types here or
+/// include this file, please make sure whether you want to couple with dsn_runtime or not.
+
 #pragma once
 
 #include <dsn/tool-api/task_spec.h>
@@ -194,21 +198,6 @@ CONFIG_FLD_STRING(rwlock_nr_factory_name, "", "non-recurisve rwlock provider")
 CONFIG_FLD_STRING(semaphore_factory_name, "", "semaphore provider")
 CONFIG_FLD_STRING(logging_factory_name, "", "logging provider")
 CONFIG_END
-
-enum sys_exit_type
-{
-    SYS_EXIT_NORMAL,
-    SYS_EXIT_BREAK, // Ctrl-C/Break,Shutdown,LogOff, see SetConsoleCtrlHandler
-    SYS_EXIT_EXCEPTION,
-
-    SYS_EXIT_INVALID
-};
-
-ENUM_BEGIN(sys_exit_type, SYS_EXIT_INVALID)
-ENUM_REG(SYS_EXIT_NORMAL)
-ENUM_REG(SYS_EXIT_BREAK)
-ENUM_REG(SYS_EXIT_EXCEPTION)
-ENUM_END(sys_exit_type)
 
 extern const char *FLAGS_aio_factory_name;
 } // namespace dsn
