+++ src/core/core/logging.cpp
@@ -24,11 +24,8 @@
  * THE SOFTWARE.
  */
 
-#include <dsn/service_api_c.h>
 #include <dsn/tool-api/command_manager.h>
 #include <dsn/tool-api/logging_provider.h>
-#include <dsn/tool_api.h>
-#include "service_engine.h"
 #include <dsn/tool-api/auto_codes.h>
 #include <dsn/utility/flags.h>
 #include <dsn/utility/smart_pointers.h>
